Add files via upload
This is Task-1 of my Web Development and Designing internship at Oasis Infobyte that is creating a landing page. This landing page on the topic of Ethical Hacking is built using HTML code and includes some images to make it more visually appealing. The images used on the page are relevant to the topic of Ethical Hacking and help to create a professional look and feel. The HTML code used to create the landing page is simple and easy to understand, making it accessible to beginners who may not have experience with web development. With a little bit of customization, you can make the page your own and tailor it to the unique needs of your business.